Moved View IO files into a properly named package.
