--- src/main/java/CoreEntities/IO/ViewController.java
@@ -1,4 +1,0 @@
-package CoreEntities.IO;
-
-public class ViewController {
-}
