+++ src/main/java/io/ui/logic/ViewPresenter.java
@@ -1,4 +1,4 @@
-package CoreEntities.IO;
+package io.ui.logic;
 
 public class ViewPresenter {
 }
